bcakend/requirements.txt
grpcio-tools
pandas
prisma
asyncio
bcrypt
pyjwt